+++ src/euroeval/prompt_templates/__init__.py
@@ -13,6 +13,7 @@ from .sentiment_classification import SENT_TEMPLATES
 from .simplification import SIMPL_TEMPLATES
 from .summarization import SUMM_TEMPLATES
 from .token_classification import TOKEN_CLASSIFICATION_TEMPLATES
+from .translation import TRANSLATION_TEMPLATES
 
 if t.TYPE_CHECKING:
     from ..languages import Language