+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added a new `translation` task. This uses BERTScore and ROUGE-L as metrics, just like
+  the `summarization` task. This was added by @oliverkinch ✨
 - Now supports the `detectable_format:constrained_response_with_argument` IFEval
   constraint, being the same as `detectable_format:constrained_response` but with
   a list of options to check for, rather than a hardcoded list of English options.